Use int64 for index & term

+++ util.go
@@ -78,10 +78,10 @@ func (b *bufferedReader) Read(p []byte) (int, error) {
 	}
 }
 
-func toLogEntries(term int, nextIndex int, commands [][]byte) []*pb.LogEntry {
+func toLogEntries(term int64, nextIndex int64, commands [][]byte) []*pb.LogEntry {
 	entries := make([]*pb.LogEntry, len(commands))
 	for i, command := range commands {
-		entries[i] = &pb.LogEntry{Term: int32(term), Index: int32(nextIndex), Command: command}
+		entries[i] = &pb.LogEntry{Term: term, Index: nextIndex, Command: command}
 		nextIndex++
 	}
 	return entries
